requirements.txt
discord.py==2.3.1
selenium==4.10.0
playwright==1.52.0
webdriver-manager==3.8.6
pyvirtualdisplay 